Update the README to achieve the project
The project is now being archieved off so we can focus our efforts
on to create an awsome Beta service.  As this is a public repo it
makes sense to update the README in case anyone stumbles over it
in the future and make it avaailable to the Welsh Government.